# shared/openapi/family-cfo.v1.yaml
openapi: 3.1.0
info:
  title: Family CFO API
  version: 0.1.0
  description: >
    Initial specification for the Family CFO self-hosted backend API.
    This contract is the source of truth for generated Swift and Angular clients.
servers:
  - url: /api/v1
    description: Local Family CFO server
...
paths:
...
  /household/members:
    get:
      tags:
        - Household
      operationId: listMembers
      summary: List household members
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Household members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Household
      operationId: createMember
      summary: Add a household member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberCreateRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /household/members/{user_id}:
    patch:
      tags:
        - Household
      operationId: updateMemberRole
      summary: Change a member's role
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleUpdateRequest"
      responses:
        "200":
          description: Member role changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Household
      operationId: deleteMember
      summary: Remove a household member
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Member removed
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /transactions:
    get:
      tags:
        - Transactions
      operationId: listTransactions
      summary: List transactions for the household
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Transactions
      operationId: createTransaction
      summary: Manually create a transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /transactions/{transaction_id}:
    patch:
      tags:
        - Transactions
      operationId: updateTransaction
      summary: Update a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdateRequest"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Transactions
      operationId: deleteTransaction
      summary: Delete a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Transaction deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Money:
      type: object
      required:
        - amount_minor
        - currency
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Signed integer amount in minor currency units.
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
...
    HouseholdRole:
      type: string
      enum:
        - owner
        - adult
        - viewer
        - child
...
    Transaction:
      type: object
      required:
        - id
        - account_id
        - occurred_at
        - amount
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date
        amount:
          $ref: "#/components/schemas/Money"
        merchant:
          type: string
        category:
          type: string
        description:
          type: string
...
    TransactionListResponse:
      type: object
      required:
        - transactions
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
...
    Member:
      type: object
      required:
        - user_id
        - email
        - display_name
        - role
        - created_at
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/HouseholdRole"
        created_at:
          type: string
          format: date-time
    MemberListResponse:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    MemberCreateRequest:
      type: object
      required:
        - email
        - password
        - display_name
        - role
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/HouseholdRole"
    MemberRoleUpdateRequest:
      type: object
      required:
        - role
      properties:
        role:
          $ref: "#/components/schemas/HouseholdRole"
...
    TransactionCreateRequest:
      type: object
      required:
        - account_id
        - occurred_at
        - amount
      properties:
        account_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date
        amount:
          $ref: "#/components/schemas/Money"
        merchant:
          type: string
        description:
          type: string
    TransactionUpdateRequest:
      type: object
      properties:
        account_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date
        amount:
          $ref: "#/components/schemas/Money"
        merchant:
          type: string
        description:
          type: string
...